add Keycode type and CustomType for abstract Types
Was not able to split the commit into two

Keycode type is simply a Uint8
(and Button type, too! But Button type is outcommented because still unused)

add also a CustomType class for defining complicated/unusual types
(E.G.: The Screen package gets a field named "backing-stores" which can
be
0 for Never
1 for WhenMapped
2 for Always
this can be represented like this:

CustomType.new() do |never, when_mapped, always|
never = 0
when_mapped = 1
always = 2
end

The instance variables will be set automatically. The object just needs
to be interpreted the right way.

I'm not sure if I understood the whole thing, but if, this helps to
define types which can have just x values (as in the example).